bot/requirements.txt
discord.py>=1.6
asyncpg
emoji
pycountry
beautifulsoup4
lxml